+++ CHANGELOG.md
@@ -22,6 +22,10 @@ all APIs might be changed.
 
 - Handle the breaking change to `YamlElixir.read_from_file` in YamlElixir 1.4.0
 
+### Bug Fixes
+
+- Token & certificate auth details can now be supplied in the mix config.
+
 ## v0.6.0 - 2018-02-19
 
 ### Breaking Changes